add comando de deletar tabela

+++ filme/management/commands/resetar_tabela.py
@@ -4,9 +4,12 @@ from filme.models import Filme
 class Command(BaseCommand):
     help = 'Deleta todos dados da tabala filme'
     
-    def handle(self):
-        try:
+    def add_arguments(self, parser):
+        parser.add_argument('--drop-all', action='store_true', help=help)
+        
+    def handle(self, *args, **options):
+        if options.get("drop_all"):
+            self.warn("Apangando registros")
             Filme.objects.all().delete()
         
-        except:
-            raise CommandError('Erro ao deletar tabela')
+
